+++ examples/local/main.tf
@@ -11,15 +11,17 @@ provider "terrahelm" {
   kube_context = "rancher-desktop"
 }
 
-resource "terrahelm_release" "nginx" {
-  name             = "nginx"
-  git_repository   = "https://github.com/bitnami/charts.git"
-  git_reference    = "main"
-  chart_path       = "bitnami/nginx"
-  namespace        = "nginx"
-  create_namespace = true
-  timeout          = 60
-  atomic           = true
+#
+# git_repository
+#
+resource "terrahelm_release" "git_repository" {
+  name           = "example"
+  git_repository = "https://github.com/helm/examples.git"
+  git_reference  = "main"
+  chart_path     = "charts/hello-world"
+  timeout        = 60
+  atomic         = true
+  insecure       = true
 
   values = <<EOF
   replicaCount: 1
@@ -28,8 +30,30 @@ resource "terrahelm_release" "nginx" {
   EOF
 }
 
+#
+# chart_url
+#
+resource "terrahelm_release" "chart_url_http" {
+  name             = "traefik"
+  chart_url        = "https://traefik.github.io/charts/traefik/traefik-26.1.0.tgz//traefik"
+  namespace        = "traefik"
+  create_namespace = true
+  timeout          = 60
+  atomic           = true
+}
+
+resource "terrahelm_release" "chart_url" {
+  name             = "nginx"
+  chart_url        = "github.com/kubernetes/ingress-nginx//charts/ingress-nginx?ref=helm-chart-4.8.3&depth=1"
+  namespace        = "nginx"
+  create_namespace = true
+  timeout          = 60
+  atomic           = true
+}
+
+
 output "release_status" {
-  value = terrahelm_release.nginx.release_status
+  value = terrahelm_release.git_repository.release_status
 }
 
 data "terrahelm_release" "nginx" {
@@ -37,7 +61,7 @@ data "terrahelm_release" "nginx" {
   namespace = "nginx"
 
   depends_on = [
-    terrahelm_release.nginx
+    terrahelm_release.chart_url
   ]
 }
 
@@ -45,9 +69,12 @@ output "data_nginx" {
   value = data.terrahelm_release.nginx
 }
 
-resource "terrahelm_release" "fluentd" {
+#
+# chart_repository
+#
+resource "terrahelm_release" "chart_repository" {
   name             = "fluentd"
-  helm_repository  = "bitnami"
+  chart_repository = "bitnami"
   chart_path       = "fluentd"
   namespace        = "fluentd"
   create_namespace = true
@@ -59,10 +86,33 @@ data "terrahelm_release" "fluentd" {
   namespace = "fluentd"
 
   depends_on = [
-    terrahelm_release.fluentd
+    terrahelm_release.chart_repository
   ]
 }
 
 output "data_fluentd" {
   value = data.terrahelm_release.fluentd
+}
+
+#
+# values_files
+#
+resource "terrahelm_release" "chart_values_files" {
+  name             = "nginx-values"
+  chart_url        = "github.com/mikhae1/terraform-provider-terrahelm//tests/charts/?ref=master&depth=1"
+  chart_path       = "./nginx"
+  namespace        = "nginx-values"
+  create_namespace = true
+  timeout          = 60
+  atomic           = true
+
+  values_files = [
+    "https://raw.githubusercontent.com/mikhae1/terraform-provider-terrahelm/master/tests/charts/values/nginx/common.yaml",
+    "./values/nginx/dev-values.yaml",
+  ]
+
+  values = <<EOF
+  serviceAccount:
+    name: overriden
+  EOF
 }